Mixer works but take the Class AufioOutputI2Sesp32
Mixer In and Mixer Out works fine now. In the Arduino-Sketch, please use the AudioOutput2Sesp32 rather than the new AudioOutputI2S! The result with the special version which was based on an older version is better for the usage in this case. AudioMixerInBuffer and AudioMixerOutBuffer work well and if all samples stopped to play, the DAC will be shutdown without extra noise.