Mixer works fine
This Test-Sketch could be used to create Your own sample-based e-drum, e-percussion or instrument or use it as the startingpoint for your sound-art-installation.
It uses the library https://github.com/earlephilhower/ESP8266Audio and You have to extend this library by adding some additional files to it.
It is able to play 4 samples out of progmem at the same time and repeats them for a while.
You could add buttons, touch sensors or piezos to this project to create Your own instrument.
The class AudioGeneratorSAMPLE will be extended it the next time to be able to add some effects to the sample. 
Finally, check out this project and the project of Earlee to get the latest news around this sample-player and media-player in general based on the espressif ESP32.  

The samples are 44.1kHz samplerate and 16Bit. Converted with a local converter "hex2h.html" into h-files.

Used Hardware:
- ESP32 Devboard
- MAX 98357A DAC (Adafruit)
- loadspeaker 8Ohm

Used Software:
- Library: https://github.com/earlephilhower/ESP8266Audio
- my additional classes
- Arduino Development Environment

For testing, my family advised me to use a switch to activate the DAC only when needed. .. so I used a pushbutton between the Vin of the DAC and the 5Volt-Line of the Breadboard. This way, it is less annoying for everyone. :)

Next steps:
- adding effects to the AudioGeneratorSAMPLE
- perhaps replacing the Class AudioOutputI2Sesp32 with the "official" AudioOutput from the master branch maintened by Earlee Phil Hower. 
- testing with more channels in the mixer .. 16 are planned.