Create esp8266_piezo_Multiread.ino
This file uses WAV-Files which I converted into "h"-Files. It uses the Audioesp8266-Library in a version from december 2017. The main test was done in the if-then-clause starting with line 151. The trick here was to find a way to get a precise MAX of the Piezo.
Testpurpose was to get a valid Velocity (MAX) and a running simple display.