Create OLED_UI.ino
The test here, I had to test the Capabilities for the menu. In the Prototype1 I only used the OLED to display the name of the static preset. When going over to manage dynamic presets, I have to provide a simple GUI which is presented on a OLED (I2S-Type) with 128x64 pixels on 0.96".
I have planned to use 2 Rotary-Encoder which have integrated pushbuttons to create a complete menu. Most influence to the UI came from the stupid UIs and workflows from modern TVs and theri limited capabilities of the IR-Remote.