Mixer works only with AudioSourcePROGMEM!!!
After many tests I came to the final, the Mixer works only correct with the AudioSourcePROGMEM. All Tests to get it running with AudioSourceSD are not well. It sounds like Ringmodulation and the Play-Speed is lower if a Audio-Source-SD and another Audio-Source  are active. I give up  with the mixer at this point.